Automation update
## Even easier to automate with cron!

This is a major overhaul of how config files are loaded and includes some bugfixes where the database usable check would succeed even if the database didn't exist.

Config files are now loaded by default from the conf.d directory in the repo folder and -c -config can either specify a directory containing config files or a specific config file. If a directory is specified the script will use each config file, compare the version number and then run the checks and backup. Passing the -t argument will result in all config files loaded to be dry-run so no changes are made.

Added .conf files to gitignore and the template config file is now just called template inside the conf.d folder and as such will not be read by the script so it can be safely left.